UI Improvements
Changed the new and history buttons to have an ellipse with the light tan background on selected instead of a white background.
Added 4 images on the main screen.
Changed main screen to be full screen.
Added an info button instead of a context menu About button due to being a full screen app.